+++ src/types/registration.ts
@@ -1,10 +1,8 @@
-import { APIError } from '@okta/okta-auth-js';
+import { APIError, Input } from '@okta/okta-auth-js';
+import type { IdxOption } from '@okta/okta-auth-js/types/lib/idx/types/idx-js';
 import {
   SimpleCallback
 } from './results';
-import {
-  RegistrationElementSchema
-} from '../v3/src/types';
 
 // Registration
 
@@ -101,6 +99,18 @@ interface RegistrationSchemaV1 {
     fieldOrder: Array<string>;
   };
 }
+
+// Utility to modify interfaces / types
+type Modify<T, R> = Omit<T, keyof R> & R;
+
+export type RegistrationElementSchema = Modify<Input, {
+  'label-top'?: boolean;
+  placeholder?: string;
+  'data-se'?: string;
+  options?: IdxOption[] | Record<string, string>;
+  sublabel?: string;
+  wide?: boolean;
+}>;
 
 export type RegistrationSchema = RegistrationSchemaV1 | RegistrationElementSchema[];
